Minor refactoring; removed zipkin URL
