+++ zipkin-hadoop-job-runner/src/main/resources/email.mustache
@@ -37,7 +37,7 @@
 
             {{#tableUrlRows}}
             <tr>
-            <td><a href="https://zipkin.smf1.twitter.com/traces/{{urlToken1}}">{{urlToken2}}</td>
+            <td><a href="{{urlToken1}}">{{urlToken2}}</td>
             {{#tableUrlRowTokens}}
             <td>{{tableUrlRowToken}}</td>
             {{/tableUrlRowTokens}}
