+++ zipkin-hadoop-job-runner/src/main/scala/com/twitter/zipkin/hadoop/HadoopJobClient.scala
@@ -86,7 +86,6 @@ abstract class HadoopJobClient(val combineSimilarNames: Boolean) {
     Util.traverseFileTree(processFile, file)
     for (t <- serviceToValues) {
       val (service, values) = t
-      println(service)
       processKey(service, values)
     }
   }
